# internal/assets/openapi.yaml
openapi: 3.0.3
info:
  title: File Management API
  description: API for managing files, folders, tags, and semantic search
  version: 1.0.0
  contact:
    name: RxTech Lab
    url: https://github.com/rxtech-lab/files-management
servers:
  - url: /api
    description: API server
...
paths:
...
  /api/tags:
    get:
      tags:
        - Tags
      summary: List tags
      description: Returns a paginated list of tags with optional search
      operationId: listTags
      parameters:
        - name: keyword
          in: query
          description: Search keyword for tag name
          schema:
            type: string
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of tags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Tags
      summary: Create tag
      description: Creates a new tag
      operationId: createTag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTagRequest'
      responses:
        '201':
          description: Tag created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/tags/{id}:
    get:
      tags:
        - Tags
      summary: Get tag
      description: Returns a tag by ID
      operationId: getTag
      parameters:
        - $ref: '#/components/parameters/TagId'
      responses:
        '200':
          description: Tag details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Tags
      summary: Update tag
      description: Updates an existing tag
      operationId: updateTag
      parameters:
        - $ref: '#/components/parameters/TagId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTagRequest'
      responses:
        '200':
          description: Tag updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Tags
      summary: Delete tag
      description: Deletes a tag
      operationId: deleteTag
      parameters:
        - $ref: '#/components/parameters/TagId'
      responses:
        '204':
          description: Tag deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Folders
...
  /api/upload:
    post:
      tags:
        - Upload
      summary: Upload file
      description: Uploads a file to S3-compatible storage
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
      responses:
        '201':
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/upload/presigned:
    get:
      tags:
        - Upload
      summary: Get presigned upload URL
      description: Generates a presigned URL for direct upload to S3
      operationId: getPresignedURL
      parameters:
        - name: filename
          in: query
          required: true
          description: Name of the file to upload
          schema:
            type: string
        - name: content_type
          in: query
          description: MIME type of the file
          schema:
            type: string
            default: application/octet-stream
      responses:
        '200':
          description: Presigned URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedURLResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  parameters:
    TagId:
      name: id
      in: path
      required: true
      description: Tag ID
      schema:
        type: integer
...
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message

    # Tags
    Tag:
      type: object
      required:
        - id
        - user_id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        user_id:
          type: string
        name:
          type: string
        color:
          type: string
          description: Hex color code (e.g., #FF5733)
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTagRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        color:
          type: string
        description:
          type: string
    UpdateTagRequest:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
        description:
          type: string
    TagListResponse:
      type: object
      required:
        - data
        - total
        - limit
        - offset
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    UploadResponse:
      type: object
      required:
        - key
        - filename
        - size
        - content_type
      properties:
        key:
          type: string
          description: S3 object key
        download_url:
          type: string
          format: uri
        filename:
          type: string
        size:
          type: integer
        content_type:
          type: string
    PresignedURLResponse:
      type: object
      required:
        - upload_url
        - key
        - content_type
      properties:
        upload_url:
          type: string
          format: uri
        key:
          type: string
        content_type:
          type: string
security:
  - BearerAuth: []